requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

httpx==0.27.2

pandas==2.2.3
numpy==1.26.4
scipy==1.13.1
ta==0.11.0

xgboost==2.1.1

google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.27.0
db-dtypes==1.3.0
pyarrow==17.0.0
